cleanup inits
